#1479 - change TAB_MAP to a class property
